added an equals method in chat object which compares 2 objects according to uid.
in the data change of all chats activity - changed the notify data item inserted to data item changed according to condition if it already exists or a new one is added.

the code to update the last message in specific chat was moved fromm on child added to sendMessage so that the database is updated only once.

managed to correct the error in no of users, there was a typo in it.