Restructured the whole code. Cannot mention all the details but created allContacts which will create single chat with all the users present in the user's contact list and will store the the in hashMap for future reference.
Another major change was the storage of hashMap and user in a file which will be extracted every time the app is started and hence reduce time to take details from database every time.
Also some major code changes in all chats activity on how the message data is accessed.
Added OfflineApplication for set Persistence enabled so that it is called only once
And also the minor changes to support the major changes made.